test: 각 도메인 별로 동시성 테스트의 경우 파일 분리
